requirements.txt
Django==1.10.3
psycopg2==2.6.2
dj-database-url==0.4.1
djangorestframework==3.4.1
drf-nested-routers==0.11.1



